types refactoring, updated tsconfig to support ES2022 to support Object.hasOwn, added vitest which introduces unit tests, updated prettier config, introduced parseSettings() and loadSettings()
