+++ .prettierrc
@@ -1,6 +1,8 @@
 {
+  "semi": true,
+  "singleQuote": false,
   "trailingComma": "es5",
   "tabWidth": 2,
-  "printWidth": 180,
+  "printWidth": 120,
   "endOfLine": "auto"
 }
