+++ package.json
@@ -32,6 +32,7 @@
     "prettier": "3.6.2",
     "typescript": "~5.8.3",
     "typescript-eslint": "^8.44.1",
-    "vite": "^6.3.6"
+    "vite": "^6.3.6",
+    "vitest": "^3.2.4"
   }
 }
