--- src/utilities.ts
@@ -1,4 +1,0 @@
-export function lazyToggleTheme() {
-  const htmlElement = document.documentElement;
-  htmlElement.style.filter = htmlElement.style.filter !== "invert(1)" ? "invert(1)" : "invert(0)";
-}
